Fix schedule_cycle_sum overflow in sdiag
Changelog: Change schedule_cycle_sum to uint64_t to prevent overflows which
 will otherwise occur after 2^32 microseconds (71.6 minutes).
Ticket: 23874

+++ slurm/slurm.h
@@ -3320,7 +3320,7 @@ typedef struct stats_info_response_msg {
 
 	uint32_t schedule_cycle_max;
 	uint32_t schedule_cycle_last;
-	uint32_t schedule_cycle_sum;
+	uint64_t schedule_cycle_sum;
 	uint32_t schedule_cycle_counter;
 	uint32_t schedule_cycle_depth;
 	uint32_t *schedule_exit;
